Constrain external camera config
The external camera should be a lower quality camera without feature
support for many of the features on the back or front camera configs.

Specifically, the external camera should have EXTERNAL capabilities, 1080p resolution, and exclude all advanced camera features. It should mimic a cheap USB webcam as closely as possible.

This is required because TV does not announce support for fully-featured cameras, and does not load many of the required config files, to correctly emulate an actual TV Device. The mismatch between the camera config and the TV device declarations is causing many test failures.

Fixes: 302126138
Fixes: 303789087
Fixes: 303957446
Fixes: 288415316
Test: Run CtsCameraTestCases and SystemFeaturesTest.
Change-Id: I48b6891168fde1435cfdedb9f564b67155deb0a5