Fix C++23 build: remove nullptr initializer
The initializer isn't necessary to ensure that the field is
initialized, because std::unique_ptr has a default constructor, and
this `= nullptr` syntax triggers Clang to instantiate
`std::unique_ptr<ResultProcessor>::~unique_ptr()` (when building for
C++23), which doesn't work because ResultProcessor is an incomplete
type.

Omitting the initializer is more idiomatic anyway, so remove it.

Bug: none
Test: m libgooglecamerahalutils
Flag: EXEMPT refactor
Change-Id: If15779bd6c72d8710f5d09234167ff5fb52ae4e9